requirements.txt
numpy
matplotlib
pyyaml
casadi
torch
gymnasium
scipy
